+++ packages/service-core/src/storage/BucketStorage.ts
@@ -79,8 +79,7 @@ export interface BucketStorageFactory extends AsyncDisposableObserverClient<Buck
    * Update sync rules from configuration, if changed.
    */
   configureSyncRules(
-    sync_rules: string,
-    options?: { lock?: boolean }
+    options: UpdateSyncRulesOptions
   ): Promise<{ updated: boolean; persisted_sync_rules?: PersistedSyncRulesContent; lock?: ReplicationLock }>;
 
   /**
@@ -90,6 +89,8 @@ export interface BucketStorageFactory extends AsyncDisposableObserverClient<Buck
 
   /**
    * Deploy new sync rules.
+   *
+   * Similar to configureSyncRules, but applies the update unconditionally.
    */
   updateSyncRules(options: UpdateSyncRulesOptions): Promise<PersistedSyncRulesContent>;
 
@@ -232,6 +233,7 @@ export interface PersistedSyncRules {
 export interface UpdateSyncRulesOptions {
   content: string;
   lock?: boolean;
+  validate?: boolean;
 }
 
 export interface SyncRulesBucketStorageOptions {