+++ test-client/CHANGELOG.md
@@ -1,5 +1,20 @@
 # test-client
 
+## 0.1.25
+
+### Patch Changes
+
+- Updated dependencies [ffc8d98]
+  - @powersync/service-core@0.18.1
+
+## 0.1.24
+
+### Patch Changes
+
+- Updated dependencies [4b43cdb]
+- Updated dependencies [9a9e668]
+  - @powersync/service-core@0.18.0
+
 ## 0.1.23
 
 ### Patch Changes